+++ docs/releases.rst
@@ -43,6 +43,14 @@ issues simpler.
 
 Normalized JSON export was dropped. Documentation has been stripped of techie jargon.
 
+20.7
+---------------
+
+Features already present in the 20.7 branch:
+
+* It is now possible to have controlled writers without IPI name numbers and PR affiliations.
+* Data import
+
 20.7 and beyond
 ---------------